+++ acl-groovy-dsl/src/main/groovy/javasabr/mqtt/acl/groovy/dsl/loader/AclRulesLoader.groovy
@@ -1,9 +1,9 @@
-package javasabr.mqtt.service.acl
+package javasabr.mqtt.acl.groovy.dsl.loader
 
 import javasabr.mqtt.acl.engine.exception.AclConfigurationException
 import javasabr.mqtt.acl.engine.model.rule.Rule
 import javasabr.mqtt.model.acl.Operation
-import javasabr.mqtt.service.acl.builder.AclRulesBuilder
+import javasabr.mqtt.acl.groovy.dsl.builder.AclRulesBuilder
 import javasabr.mqtt.service.acl.builder.RuleContainerBuilder
 import javasabr.rlib.collections.array.Array
 import org.codehaus.groovy.control.CompilerConfiguration
